% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SkeletonDistributedModel.R
\docType{package}
\name{SkeletonDistributedModel}
\alias{SkeletonDistributedModel}
\title{SkeletonDistributedModel}
\description{
A package for running distributed model studies using data in the OMOP CDM
}
